+++ crates/database/interface/src/lib.rs
@@ -12,10 +12,12 @@ use state::{Account, AccountInfo, Bytecode};
 #[cfg(feature = "asyncdb")]
 pub mod async_db;
 pub mod empty_db;
+pub mod try_commit;
 
 #[cfg(feature = "asyncdb")]
 pub use async_db::{DatabaseAsync, WrapDatabaseAsync};
 pub use empty_db::{EmptyDB, EmptyDBTyped};
+pub use try_commit::{ArcUpgradeError, TryDatabaseCommit};
 
 /// EVM database interface.
 #[auto_impl(&mut, Box)]
@@ -41,63 +43,6 @@ pub trait Database {
 pub trait DatabaseCommit {
     /// Commit changes to the database.
     fn commit(&mut self, changes: HashMap<Address, Account>);
-}
-
-/// EVM database commit interface that can fail.
-///
-/// This is intended for use with types that may fail to commit changes, e.g.
-/// because they are directly interacting with the filesystem, or must arrange
-/// access to a shared resource.
-pub trait TryDatabaseCommit {
-    /// Error type for when [`TryDatabaseCommit::try_commit`] fails.
-    type Error: core::error::Error;
-
-    /// Attempt to commit changes to the database.
-    fn try_commit(&mut self, changes: HashMap<Address, Account>) -> Result<(), Self::Error>;
-}
-
-impl<Db> TryDatabaseCommit for Db
-where
-    Db: DatabaseCommit,
-{
-    type Error = core::convert::Infallible;
-
-    #[inline]
-    fn try_commit(&mut self, changes: HashMap<Address, Account>) -> Result<(), Self::Error> {
-        self.commit(changes);
-        Ok(())
-    }
-}
-
-#[cfg(feature = "std")]
-/// Error type for implementation of [`TryDatabaseCommit`] on
-/// [`std::sync::Arc`].
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct ArcUpgradeError;
-
-#[cfg(feature = "std")]
-impl core::fmt::Display for ArcUpgradeError {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        write!(f, "Arc reference is not unique, cannot mutate")
-    }
-}
-
-#[cfg(feature = "std")]
-impl core::error::Error for ArcUpgradeError {}
-
-#[cfg(feature = "std")]
-impl<Db> TryDatabaseCommit for std::sync::Arc<Db>
-where
-    Db: DatabaseCommit + Send + Sync,
-{
-    type Error = ArcUpgradeError;
-
-    #[inline]
-    fn try_commit(&mut self, changes: HashMap<Address, Account>) -> Result<(), Self::Error> {
-        std::sync::Arc::get_mut(self)
-            .map(|db| db.commit(changes))
-            .ok_or(ArcUpgradeError)
-    }
 }
 
 /// EVM database interface.
